fixup! fixup! Use type-generic or float versions of cmath functions where appropriate.

+++ gnuradio-runtime/include/gnuradio/fxpt.h
@@ -53,7 +53,7 @@ namespace gr {
       float_to_fixed(float x)
     {
       // Fold x into -PI to PI.
-      int d = (int)floorf(x/TAU+0.5);
+      int d = (int)std::floor(x/TAU+0.5);
       x -= d*TAU;
       // And convert to an integer.
       return (int32_t) ((float) x * TWO_TO_THE_31 / PI);
